# Faster IndexOf for arm64