@dakersnar 2022-04-21
Performance improvement on Windows arm64: https://github.com/dotnet/perf-autofiling-issues/issues/4732